refactor: 기존 이벤트 페이지 구조 변경
