+++ apps/web/src/app/[locale]/event/(chrismas)/Content.tsx
@@ -1,0 +1,3 @@
+export function ChristmasContent() {
+  return <div>ChristmasContent</div>;
+}
