+++ apps/web/src/app/[locale]/event/(common)/Draw.tsx
@@ -21,11 +21,13 @@ import { GIT_ANIMALS_MAIN_URL } from '@/constants/outlink';
 import { Link, useRouter } from '@/i18n/routing';
 import { trackEvent } from '@/lib/analytics';
 
-import { HalloweenCard } from './HalloweenCard';
+interface DrawProps {
+  renderCard: (type: string) => React.ReactNode;
+  bonusEventCode: string;
+  baseEventCode: string;
+}
 
-const HALLOWEEN_STAR_BONUS_EVENT_CODE = 'HALLOWEEN_2024_STAR_BONUS';
-
-export const Draw = wrap.Suspense().on(() => {
+export const Draw = wrap.Suspense().on(({ renderCard, bonusEventCode, baseEventCode }: DrawProps) => {
   const t = useTranslations('Event.Halloween');
 
   const { eventCode } = useParams();
@@ -43,7 +45,7 @@ export const Draw = wrap.Suspense().on(() => {
 
   const onClickDraw = () => {
     if (!session) {
-      login('/event/HALLOWEEN_2024');
+      login(`/event/${baseEventCode}`);
       trackEvent(`${upperCaseEventCode}-not-signed-in`, {
         coupon: upperCaseEventCode,
       });
@@ -81,7 +83,7 @@ export const Draw = wrap.Suspense().on(() => {
   };
   useOutsideClick(modalRef, onClickOutside);
 
-  const isStarBonusEvent = eventCode === HALLOWEEN_STAR_BONUS_EVENT_CODE;
+  const isStarBonusEvent = eventCode === bonusEventCode;
 
   const drawButtonText = (() => {
     if (!session) {
@@ -116,13 +118,13 @@ export const Draw = wrap.Suspense().on(() => {
           <div className={fixedStyle} ref={modalRef}>
             <motion.div className={modalStyle} variants={modalVariants} initial="initial" animate="animate" exit="exit">
               <span className={resultTitleStyle}>{t('result-title')}</span>
-              <HalloweenCard type={drawedPet} />
+              {renderCard(drawedPet)}
 
               <Link href="/mypage" className={resultAnchorStyle}>
                 {t('result-apply')}
               </Link>
 
-              <StarAnchor />
+              <StarAnchor bonusEventCode={bonusEventCode} />
             </motion.div>
           </div>
         )}
@@ -184,7 +186,7 @@ const resultAnchorStyle = css({
 const StarAnchor = wrap
   .ErrorBoundary({ fallback: null })
   .Suspense()
-  .on(() => {
+  .on(({ bonusEventCode }: { bonusEventCode: string }) => {
     const t = useTranslations('Event.Halloween');
     const { data: session } = useSession();
     if (!session?.user.name) {
@@ -196,11 +198,11 @@ const StarAnchor = wrap
     } = useSuspenseQuery(renderQueries.isPressStar({ login: session.user.name }));
 
     const { eventCode } = useParams();
-    const isStarBonusEvent = eventCode === HALLOWEEN_STAR_BONUS_EVENT_CODE;
+    const isStarBonusEvent = eventCode === bonusEventCode;
 
     const router = useRouter();
     const onClickFirstEvent = () => {
-      router.replace(`/event/${HALLOWEEN_STAR_BONUS_EVENT_CODE}`);
+      router.replace(`/event/${bonusEventCode}`);
     };
 
     if (isStarBonusEvent) {
@@ -210,7 +212,7 @@ const StarAnchor = wrap
     if (isPressStar) {
       return (
         <Button>
-          <Link href={`/event/${HALLOWEEN_STAR_BONUS_EVENT_CODE}`}>{t('result-already-star')}</Link>
+          <Link href={`/event/${bonusEventCode}`}>{t('result-already-star')}</Link>
         </Button>
       );
     }
