+++ apps/web/src/app/[locale]/event/(halloween)/HalloweenCard.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React from 'react';
 import Image from 'next/image';
 import { css } from '_panda/css';
