+++ apps/web/src/app/[locale]/event/(halloween)/KingGhost.tsx
