+++ apps/web/src/app/[locale]/event/[eventCode]/layout.tsx
@@ -1,102 +1,29 @@
 import React from 'react';
-import type { Metadata } from 'next';
-import Image from 'next/image';
-import { css, cx } from '_panda/css';
 
 import GNB from '@/components/GNB/GNB';
 
 import { Footer } from '../../landing/Footer';
 
-import { KingGhost } from './KingGhost';
+export async function generateMetadata({ params }: { params: { eventCode: string } }) {
+  return {
+    openGraph: {
+      images: [
+        {
+          url: `/event/${params.eventCode}/og-image.png`,
+        },
+      ],
+    },
+  };
+}
 
-export const metadata: Metadata = {
-  openGraph: {
-    images: [
-      {
-        url: '/og-image-event.png',
-      },
-    ],
-  },
-};
-
-function HalloweenEventLayout({ children }: { children: React.ReactNode }) {
+function EventLayout({ children }: { children: React.ReactNode }) {
   return (
     <div>
       <GNB />
-      <div className={bgContainerStyle}>
-        <Image
-          src="/event/halloween/halloween-bg.webp"
-          alt="halloween bg"
-          layout="fill"
-          objectFit="cover"
-          className={bgImageStyle}
-        />
-        <Image
-          src="/event/halloween/halloween-right.webp"
-          alt="halloween bg"
-          width={544}
-          height={1470}
-          objectFit="contain"
-          className={cx(imageStyle, rightImageStyle)}
-        />
-        <Image
-          src="/event/halloween/halloween-left.webp"
-          alt="halloween bg"
-          width={561}
-          height={1470}
-          objectFit="contain"
-          className={cx(imageStyle, leftImageStyle)}
-        />
-
-        <div className={showMobile}>
-          <KingGhost />
-        </div>
-
-        {children}
-      </div>
+      {children}
       <Footer />
     </div>
   );
 }
 
-export default HalloweenEventLayout;
-
-const bgImageStyle = css({
-  pointerEvents: 'none',
-});
-const bgContainerStyle = css({
-  position: 'relative',
-  width: '100%',
-  minHeight: 'calc(100vh - 60px)',
-  fontFamily: 'Product Sans',
-});
-
-const imageStyle = css({
-  width: 'auto',
-  height: '100%',
-  position: 'absolute',
-  objectFit: 'contain',
-  zIndex: 1,
-  minHeight: 'calc(100vh - 60px)',
-  pointerEvents: 'none',
-  top: '-60px',
-  _mobile: {
-    display: 'none',
-  },
-});
-
-const rightImageStyle = css({
-  right: 0,
-});
-
-const leftImageStyle = css({
-  position: 'absolute',
-  left: 0,
-});
-
-const showMobile = css({
-  display: 'none',
-  _mobile: {
-    display: 'block',
-  },
-});
+export default EventLayout;
